Implemented solar display for rgb digits. Working, but super inefficient with reconstruction of references that should be in the struct
