+++ Cargo.toml
@@ -22,6 +22,7 @@ tryhard = "0.5.1"
 
 ws2818-rgb-led-spi-driver = { version = "2.0.0", optional = true }
 rand = "0.8.5"
+colorgrad = "0.6.2"
 
 [features]
 i2c_display = ["dep:embedded-graphics", "dep:linux-embedded-hal", "dep:ssd1306", "dep:tinybmp", "dep:display-interface", "dep:ws2818-rgb-led-spi-driver"]
